pyproject.toml
[tool.poetry]
name = "mcp-netops-poc"
version = "0.1.0"
description = "MCP gateway that exposes Model‑Context‑Protocol endpoints for Cisco network troubleshooting PoC"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mcp_gateway"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = {version = "^0.29", extras = ["standard"]}
pydantic = "^2.7"
python-dotenv = "^1.0"
netmiko = "^4.2"
pyats = {version = "^23.11", extras = ["genie"]}
loguru = "^0.7"
python-jose = "^3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
black = "^24.2"
isort = "^5.13"
mypy = "^1.9"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

